# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "UniTracker API",
        "description": "L'API UniTracker permet de suivre en temps réel les bus universitaires, planifier les trajets, et recevoir des notifications intelligentes. Elle facilite la gestion des transports pour les étudiants et les chauffeurs via des requêtes HTTP sécurisées.",
        "termsOfService": "https://unitracker.ucbc.cd/terms",
        "contact": {
            "name": "Équipe Support UniTracker",
            "email": "support@unitracker.cd"
        },
        "license": {
            "name": "UCBC / UniTracker",
            "url": "https://unitracker.ucbc.cd/license"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://unitracker.ucbc.cd",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/api/bus-journals": {
            "get": {
                "tags": [
                    "BusJournal"
                ],
                "summary": "Lister les trajets du carnet de bord",
                "operationId": "a016430eb2360c7d54ee7d4fed91f360",
                "parameters": [
                    {
                        "name": "bus_id",
                        "in": "query",
                        "description": "Filtrer par ID de bus",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "chauffeur_id",
                        "in": "query",
                        "description": "Filtrer par ID de chauffeur",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste des journaux",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "journaux": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "bus_id": {
                                                        "type": "integer"
                                                    },
                                                    "chauffeur_id": {
                                                        "type": "integer"
                                                    },
                                                    "depart": {
                                                        "type": "string"
                                                    },
                                                    "destination": {
                                                        "type": "string"
                                                    },
                                                    "statut": {
                                                        "type": "string"
                                                    },
                                                    "heure_depart": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "heure_arrivee": {
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "remarques": {
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "BusJournal"
                ],
                "summary": "Créer un nouveau trajet dans le carnet de bord",
                "operationId": "6ee60a305093583ccb3668eeb093b18b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "bus_id",
                                    "chauffeur_id",
                                    "depart",
                                    "destination",
                                    "heure_depart"
                                ],
                                "properties": {
                                    "bus_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "chauffeur_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "depart": {
                                        "type": "string",
                                        "example": "Campus UCBC"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "Faculté de Médecine"
                                    },
                                    "heure_depart": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-04-15T08:00:00Z"
                                    },
                                    "remarques": {
                                        "type": "string",
                                        "example": "Départ à l'heure"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Trajet enregistré avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "journal": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer"
                                                },
                                                "statut": {
                                                    "type": "string"
                                                },
                                                "depart": {
                                                    "type": "string"
                                                },
                                                "destination": {
                                                    "type": "string"
                                                },
                                                "heure_depart": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}